clean up warnings

+++ cpu/idt.c
@@ -17,7 +17,7 @@ void set_idt_gate(int n, uint32_t handler) {
 }
 
 void set_idt() {
-  idt_reg.base = (uint32_t) &idt;
+  idt_reg.base = (uintptr_t) &idt;
   idt_reg.limit = IDT_ENTRIES * sizeof(idt_gate_t) - 1;
 
   asm volatile("lidtl (%0)" : : "r" (&idt_reg));
