+++ cpu/isr.c
@@ -16,38 +16,38 @@ isr_t interrupt_handlers[256];
 
 // Can't do this with a loop because we need the address of the function names
 void isr_install() {
-  set_idt_gate(0, (uint32_t)isr0);
-  set_idt_gate(1, (uint32_t)isr1);
-  set_idt_gate(2, (uint32_t)isr2);
-  set_idt_gate(3, (uint32_t)isr3);
-  set_idt_gate(4, (uint32_t)isr4);
-  set_idt_gate(5, (uint32_t)isr5);
-  set_idt_gate(6, (uint32_t)isr6);
-  set_idt_gate(7, (uint32_t)isr7);
-  set_idt_gate(8, (uint32_t)isr8);
-  set_idt_gate(9, (uint32_t)isr9);
-  set_idt_gate(10, (uint32_t)isr10);
-  set_idt_gate(11, (uint32_t)isr11);
-  set_idt_gate(12, (uint32_t)isr12);
-  set_idt_gate(13, (uint32_t)isr13);
-  set_idt_gate(14, (uint32_t)isr14);
-  set_idt_gate(15, (uint32_t)isr15);
-  set_idt_gate(16, (uint32_t)isr16);
-  set_idt_gate(17, (uint32_t)isr17);
-  set_idt_gate(18, (uint32_t)isr18);
-  set_idt_gate(19, (uint32_t)isr19);
-  set_idt_gate(20, (uint32_t)isr20);
-  set_idt_gate(21, (uint32_t)isr21);
-  set_idt_gate(22, (uint32_t)isr22);
-  set_idt_gate(23, (uint32_t)isr23);
-  set_idt_gate(24, (uint32_t)isr24);
-  set_idt_gate(25, (uint32_t)isr25);
-  set_idt_gate(26, (uint32_t)isr26);
-  set_idt_gate(27, (uint32_t)isr27);
-  set_idt_gate(28, (uint32_t)isr28);
-  set_idt_gate(29, (uint32_t)isr29);
-  set_idt_gate(30, (uint32_t)isr30);
-  set_idt_gate(31, (uint32_t)isr31);
+  set_idt_gate(0, (uintptr_t)isr0);
+  set_idt_gate(1, (uintptr_t)isr1);
+  set_idt_gate(2, (uintptr_t)isr2);
+  set_idt_gate(3, (uintptr_t)isr3);
+  set_idt_gate(4, (uintptr_t)isr4);
+  set_idt_gate(5, (uintptr_t)isr5);
+  set_idt_gate(6, (uintptr_t)isr6);
+  set_idt_gate(7, (uintptr_t)isr7);
+  set_idt_gate(8, (uintptr_t)isr8);
+  set_idt_gate(9, (uintptr_t)isr9);
+  set_idt_gate(10, (uintptr_t)isr10);
+  set_idt_gate(11, (uintptr_t)isr11);
+  set_idt_gate(12, (uintptr_t)isr12);
+  set_idt_gate(13, (uintptr_t)isr13);
+  set_idt_gate(14, (uintptr_t)isr14);
+  set_idt_gate(15, (uintptr_t)isr15);
+  set_idt_gate(16, (uintptr_t)isr16);
+  set_idt_gate(17, (uintptr_t)isr17);
+  set_idt_gate(18, (uintptr_t)isr18);
+  set_idt_gate(19, (uintptr_t)isr19);
+  set_idt_gate(20, (uintptr_t)isr20);
+  set_idt_gate(21, (uintptr_t)isr21);
+  set_idt_gate(22, (uintptr_t)isr22);
+  set_idt_gate(23, (uintptr_t)isr23);
+  set_idt_gate(24, (uintptr_t)isr24);
+  set_idt_gate(25, (uintptr_t)isr25);
+  set_idt_gate(26, (uintptr_t)isr26);
+  set_idt_gate(27, (uintptr_t)isr27);
+  set_idt_gate(28, (uintptr_t)isr28);
+  set_idt_gate(29, (uintptr_t)isr29);
+  set_idt_gate(30, (uintptr_t)isr30);
+  set_idt_gate(31, (uintptr_t)isr31);
 
   // Remap the PIC
   port_byte_out(0x20, 0x11);
@@ -62,22 +62,22 @@ void isr_install() {
   port_byte_out(0xA1, 0x0);
 
   // Install the IRQs
-  set_idt_gate(32, (uint32_t)irq0);
-  set_idt_gate(33, (uint32_t)irq1);
-  set_idt_gate(34, (uint32_t)irq2);
-  set_idt_gate(35, (uint32_t)irq3);
-  set_idt_gate(36, (uint32_t)irq4);
-  set_idt_gate(37, (uint32_t)irq5);
-  set_idt_gate(38, (uint32_t)irq6);
-  set_idt_gate(39, (uint32_t)irq7);
-  set_idt_gate(40, (uint32_t)irq8);
-  set_idt_gate(41, (uint32_t)irq9);
-  set_idt_gate(42, (uint32_t)irq10);
-  set_idt_gate(43, (uint32_t)irq11);
-  set_idt_gate(44, (uint32_t)irq12);
-  set_idt_gate(45, (uint32_t)irq13);
-  set_idt_gate(46, (uint32_t)irq14);
-  set_idt_gate(47, (uint32_t)irq15);
+  set_idt_gate(32, (uintptr_t)irq0);
+  set_idt_gate(33, (uintptr_t)irq1);
+  set_idt_gate(34, (uintptr_t)irq2);
+  set_idt_gate(35, (uintptr_t)irq3);
+  set_idt_gate(36, (uintptr_t)irq4);
+  set_idt_gate(37, (uintptr_t)irq5);
+  set_idt_gate(38, (uintptr_t)irq6);
+  set_idt_gate(39, (uintptr_t)irq7);
+  set_idt_gate(40, (uintptr_t)irq8);
+  set_idt_gate(41, (uintptr_t)irq9);
+  set_idt_gate(42, (uintptr_t)irq10);
+  set_idt_gate(43, (uintptr_t)irq11);
+  set_idt_gate(44, (uintptr_t)irq12);
+  set_idt_gate(45, (uintptr_t)irq13);
+  set_idt_gate(46, (uintptr_t)irq14);
+  set_idt_gate(47, (uintptr_t)irq15);
 
   set_idt(); // Load with ASM
 }
